Put things back in package objects
Moving them out broke bincompat
