+++ modules/core/shared/src/main/scala/retry/package.scala
@@ -3,6 +3,8 @@ import cats.syntax.apply._
 import cats.syntax.functor._
 import cats.syntax.flatMap._
 
+import scala.concurrent.duration.FiniteDuration
+
 package object retry {
 
   /*
@@ -27,10 +29,124 @@ package object retry {
     (_, _) => Monad[M].pure(())
 
   /*
+   * Partially applied classes
+   */
+
+  private[retry] class RetryingOnFailuresPartiallyApplied[A] {
+    def apply[M[_]](
+        policy: RetryPolicy[M],
+        wasSuccessful: A => M[Boolean],
+        onFailure: (A, RetryDetails) => M[Unit]
+    )(
+        action: => M[A]
+    )(implicit
+        M: Monad[M],
+        S: Sleep[M]
+    ): M[A] = M.tailRecM(RetryStatus.NoRetriesYet) { status =>
+      action.flatMap { a =>
+        retryingOnFailuresImpl(policy, wasSuccessful, onFailure, status, a)
+      }
+    }
+  }
+
+  private[retry] class RetryingOnSomeErrorsPartiallyApplied[A] {
+    def apply[M[_], E](
+        policy: RetryPolicy[M],
+        isWorthRetrying: E => M[Boolean],
+        onError: (E, RetryDetails) => M[Unit]
+    )(
+        action: => M[A]
+    )(implicit
+        ME: MonadError[M, E],
+        S: Sleep[M]
+    ): M[A] = ME.tailRecM(RetryStatus.NoRetriesYet) { status =>
+      ME.attempt(action).flatMap { attempt =>
+        retryingOnSomeErrorsImpl(
+          policy,
+          isWorthRetrying,
+          onError,
+          status,
+          attempt
+        )
+      }
+    }
+  }
+
+  private[retry] class RetryingOnAllErrorsPartiallyApplied[A] {
+    def apply[M[_], E](
+        policy: RetryPolicy[M],
+        onError: (E, RetryDetails) => M[Unit]
+    )(
+        action: => M[A]
+    )(implicit
+        ME: MonadError[M, E],
+        S: Sleep[M]
+    ): M[A] =
+      retryingOnSomeErrors[A].apply[M, E](policy, _ => ME.pure(true), onError)(
+        action
+      )
+  }
+
+  private[retry] class RetryingOnFailuresAndSomeErrorsPartiallyApplied[A] {
+    def apply[M[_], E](
+        policy: RetryPolicy[M],
+        wasSuccessful: A => M[Boolean],
+        isWorthRetrying: E => M[Boolean],
+        onFailure: (A, RetryDetails) => M[Unit],
+        onError: (E, RetryDetails) => M[Unit]
+    )(
+        action: => M[A]
+    )(implicit
+        ME: MonadError[M, E],
+        S: Sleep[M]
+    ): M[A] = {
+
+      ME.tailRecM(RetryStatus.NoRetriesYet) { status =>
+        ME.attempt(action).flatMap {
+          case Right(a) =>
+            retryingOnFailuresImpl(policy, wasSuccessful, onFailure, status, a)
+          case attempt =>
+            retryingOnSomeErrorsImpl(
+              policy,
+              isWorthRetrying,
+              onError,
+              status,
+              attempt
+            )
+        }
+      }
+    }
+  }
+
+  private[retry] class RetryingOnFailuresAndAllErrorsPartiallyApplied[A] {
+    def apply[M[_], E](
+        policy: RetryPolicy[M],
+        wasSuccessful: A => M[Boolean],
+        onFailure: (A, RetryDetails) => M[Unit],
+        onError: (E, RetryDetails) => M[Unit]
+    )(
+        action: => M[A]
+    )(implicit
+        ME: MonadError[M, E],
+        S: Sleep[M]
+    ): M[A] =
+      retryingOnFailuresAndSomeErrors[A]
+        .apply[M, E](
+          policy,
+          wasSuccessful,
+          _ => ME.pure(true),
+          onFailure,
+          onError
+        )(
+          action
+        )
+  }
+
+  /*
    * Implementation
    */
 
-  private[retry] def retryingOnFailuresImpl[M[_], A](
+  private def retryingOnFailuresImpl[M[_], A](
       policy: RetryPolicy[M],
       wasSuccessful: A => M[Boolean],
       onFailure: (A, RetryDetails) => M[Unit],
@@ -59,7 +175,7 @@ package object retry {
     )
   }
 
-  private[retry] def retryingOnSomeErrorsImpl[M[_], A, E](
+  private def retryingOnSomeErrorsImpl[M[_], A, E](
       policy: RetryPolicy[M],
       isWorthRetrying: E => M[Boolean],
       onError: (E, RetryDetails) => M[Unit],
@@ -116,5 +232,16 @@ package object retry {
           currentStatus.cumulativeDelay
         )
     }
+
+  private[retry] sealed trait NextStep
+
+  private[retry] object NextStep {
+    case object GiveUp extends NextStep
+
+    final case class RetryAfterDelay(
+        delay: FiniteDuration,
+        updatedStatus: RetryStatus
+    ) extends NextStep
+  }
 
 }
